Updated Controller
I updated for Author, Book, Credit card, and Publisher, However I came across some issues concerning book when trying to run the program. It could be because I'm using older code, though I did do a merge with BDIAZ branch before. Will check.

+++ bookstore/src/main/java/app/bookstore/AuthorRepo.java
@@ -2,6 +2,6 @@ package app.bookstore;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepo extends CrudRepository<Author, Long> {
+public interface AuthorRepo extends CrudRepository<Author, Integer> {
 
 }
